+++ tiny_listener/__init__.py
@@ -2,7 +2,7 @@
 """
 
 
-__version__ = "0.0.7"
+__version__ = "0.0.8"
 
 
 from .listener import Listener, NotFound